First stab at search for (OP_RETURN) data functionality.
Look for data outputs (output with script starting with OP_RETURN
followed by a single data push) as new blocks are inserted. When data is
found, the transaction hash is stored with the data as key.

Add searchdatatransactions command that takes a single data byte array
and returns the transaction hashes of transactions carrying the data.